+++ rasa_core/policies/trainer.py
@@ -22,7 +22,8 @@ class PolicyTrainer(object):
 
     def train(self, filename=None, max_history=3,
               augmentation_factor=20, max_training_samples=None,
-              max_number_of_trackers=2000, remove_duplicates=True, **kwargs):
+              max_number_of_trackers=2000, remove_duplicates=True,
+              detect_conflicts=False, **kwargs):
         """Trains a policy on a domain using training data from a file.
 
         :param augmentation_factor: how many stories should be created by
@@ -36,6 +37,7 @@ class PolicyTrainer(object):
                                        story file parsing - `None` for unlimited
         :param remove_duplicates: remove duplicates from the training set before
                                   training
+        :param detect_conflicts: detects conflicting labels in the training set
         :param kwargs: additional arguments passed to the underlying ML trainer
                        (e.g. keras parameters)
         :return: trained policy
@@ -48,14 +50,16 @@ class PolicyTrainer(object):
                                            augmentation_factor,
                                            max_training_samples,
                                            max_number_of_trackers,
-                                           remove_duplicates)
+                                           remove_duplicates,
+                                           detect_conflicts)
 
         self.ensemble.train(training_data, self.domain, self.featurizer, **kwargs)
 
     def _prepare_training_data(self, filename, max_history, augmentation_factor,
                                max_training_samples=None,
                                max_number_of_trackers=2000,
-                               remove_duplicates=True):
+                               remove_duplicates=True,
+                               detect_conflicts=False):
         """Reads training data from file and prepares it for the training."""
 
         from rasa_core.training import extract_training_data_from_file
@@ -69,7 +73,8 @@ class PolicyTrainer(object):
                     augmentation_factor=augmentation_factor,
                     max_history=max_history,
                     remove_duplicates=remove_duplicates,
-                    max_number_of_trackers=max_number_of_trackers)
+                    max_number_of_trackers=max_number_of_trackers,
+                    detect_conflicts=detect_conflicts)
             if max_training_samples is not None:
                 training_data.limit_training_data_to(max_training_samples)
             return training_data